requirements.txt
PyYAML==6.0.2
reportlab==4.2.5
black==24.10.0
